fix: consolidate and streamline all zone handling
(as a side effect, this also fixes ZonalAllocationFailure handling)
