+++ pkg/providers/instancetype/instancetypes.go
@@ -172,7 +172,7 @@ func instanceTypeZones(sku *skewer.SKU, region string) sets.Set[string] {
 	skuZones := lo.Keys(sku.AvailabilityZones(region))
 	if hasZonalSupport(region) && len(skuZones) > 0 {
 		return sets.New(lo.Map(skuZones, func(zone string, _ int) string {
-			return fmt.Sprintf("%s-%s", region, zone)
+			return utils.MakeZone(region, zone)
 		})...)
 	}
 	return sets.New("") // empty string means non-zonal offering
